+++ kubernetes-client-api/src/main/java/io/fabric8/kubernetes/client/extended/leaderelection/resourcelock/LeaseLock.java
@@ -29,6 +29,10 @@ public class LeaseLock extends ResourceLock<Lease> {
     super(leaseNamespace, leaseName, identity);
   }
 
+  public LeaseLock(ObjectMeta meta, String identity) {
+    super(meta, identity);
+  }
+
   @Override
   protected Class<Lease> getKind() {
     return Lease.class;